requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
huggingface-hub==0.16.4
numpy==1.24.3
faiss-cpu==1.7.4
hdbscan==0.8.33
umap-learn==0.5.5
scikit-learn==1.3.2
pandas==2.1.3
pydantic==2.5.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
stopwordsiso==0.6.1
